requirements.txt
fastapi
uvicorn
argon2-cffi
cryptography
pyjwt
sqlalchemy
python-dotenv
requests
beautifulsoup4
psycopg2-binary
redis
icecream
pydantic
pydantic[email]
python-multipart
setuptools
